puzzle: day4
The approach is to find a trigger character first - "X" for part 1 and "A" for
part 2, before doing a search.

Use a tuple of tuples to represent the grid. Retrieving the x and y value is
annoying verbose, so create a macro to make the code look neater. For future
challenges, consider using a Map for grids.