puzzle: day7
Rather than going from left to right and testing each operation, like

    11 + 6 * 16 + 20 = 292

.. work backwards from the result, like

    292 - 20 / 16 - 6 = 11

This provides two easy early termination conditions: when a division operation
gives a remainder, or when a subtraction operation becomes negative.

Add concat by doing a "left shift". This example

    6 * 8 || 6 * 15 = 7290

.. worked backwards is simply

    7290 / 15 => 486
    486 pop 6 from the right => 48
    48 / 8 => 6

The early termination condition for a reversed concat (leftshift) is when the
number to be shifted does not match the last few digits of the remainder. In the
above example, you can't leftshift 486 by anything other than 6 or 86. Employ a
simple math trick to do this.